Add casts in termux.c to silence lint warning

+++ terminal-emulator/src/main/jni/termux.c
@@ -57,7 +57,7 @@ static int create_subprocess(JNIEnv* env,
     tcsetattr(ptm, TCSANOW, &tios);
 
     /** Set initial winsize. */
-    struct winsize sz = { .ws_row = rows, .ws_col = columns };
+    struct winsize sz = { .ws_row = (unsigned short) rows, .ws_col = (unsigned short) columns };
     ioctl(ptm, TIOCSWINSZ, &sz);
 
     pid_t pid = fork();
@@ -180,7 +180,7 @@ JNIEXPORT jint JNICALL Java_com_termux_terminal_JNI_createSubprocess(
 
 JNIEXPORT void JNICALL Java_com_termux_terminal_JNI_setPtyWindowSize(JNIEnv* TERMUX_UNUSED(env), jclass TERMUX_UNUSED(clazz), jint fd, jint rows, jint cols)
 {
-    struct winsize sz = { .ws_row = rows, .ws_col = cols };
+    struct winsize sz = { .ws_row = (unsigned short) rows, .ws_col = (unsigned short) cols };
     ioctl(fd, TIOCSWINSZ, &sz);
 }
